==> 06_ogle3_train_ogle4_check.ipynb <==
# Entreno en ogle-3 y veo que detecte en ogle 4

import pickle

import numpy as np

import pandas as pd

% matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns
# sns.set()

from sklearn.ensemble import RandomForestClassifier
from sklearn.neighbors import KNeighborsClassifier

from sklearn.preprocessing import StandardScaler

from libs.container import Container
from libs.display import d
from libs.experiment import KFoldExperiment, WithAnotherExperiment, roc

all_cat_types = pd.read_pickle(
    "data/ogle3_only/all_cat_types.pkl")

def o4_types(i):
    t = all_cat_types[all_cat_types.id == i].vs_type.values[0]
    if t == "":
        return 0
    if t.startswith("RRLyr-"):
        return 1
    return 2

class ResultCollector(object):
    
    def __init__(self):
        self.r = {}
    
    def prop_fp(self, result, size, train, test, force=False):
        key = size, train, test
        
        if force or key not in self.r:

            df = pd.DataFrame({
                "o3_cls": result.y_test_real,
                "o3_pred": result.predictions,
                "id": result.ids})

            total = len(df)

            df = df[(df.o3_pred == 1) & (df.o3_cls == 0)] 

            df["o4_cls"] = df.id.apply(o4_types)
            vstars = df[df.o4_cls != 0]
            rr = df[df.o4_cls == 1]

            prop_vstars = len(vstars) / float(total)
            prop_rrlyrae = len(rr) / float(total)
            self.r[key] = prop_vstars, prop_rrlyrae
        
        return self.r[key]

collector = ResultCollector()

sample = pd.read_pickle("data/ogle3_only/scaled/s20k.pkl.bz2")
sample["o4_type"] = sample.id.apply(lambda i: o4_types(i))
sample["tile"] = sample["id"].apply(lambda i: "b" + str(i)[1:4])
sample["cls"] = sample.vs_type.apply(lambda x: 0 if x == "" else 1)


no_features = [
    "id", "vs_catalog", "vs_type", 
    "ra_k", "dec_k", "tile", "cls", "o4_type"] 
X_columns = [c for c in sample.columns if c not in no_features]

grouped = sample.groupby("tile")
data_big = Container({k: grouped.get_group(k).copy() for k in grouped.groups.keys()})

del grouped, sample

sample = pd.read_pickle("data/ogle3_only/scaled/s5k.pkl.bz2")
sample["o4_type"] = sample.id.apply(lambda i: o4_types(i))
sample["tile"] = sample["id"].apply(lambda i: "b" + str(i)[1:4])
sample["cls"] = sample.vs_type.apply(lambda x: 0 if x == "" else 1)

grouped = sample.groupby("tile")
data_mid = Container({k: grouped.get_group(k).copy() for k in grouped.groups.keys()})

del grouped, sample

sample = pd.read_pickle("data/ogle3_only/scaled/s2_5k.pkl.bz2")
sample["o4_type"] = sample.id.apply(lambda i: o4_types(i))
sample["tile"] = sample["id"].apply(lambda i: "b" + str(i)[1:4])
sample["cls"] = sample.vs_type.apply(lambda x: 0 if x == "" else 1)

grouped = sample.groupby("tile")
data_small = Container({k: grouped.get_group(k).copy() for k in grouped.groups.keys()})

del grouped, sample

import joblib
cpu = joblib.cpu_count()

cls = {0:0, 1:1}

## Small

%%time
data = data_small
data_name = "small"

rf = KFoldExperiment(
    clf=RandomForestClassifier(n_estimators=500, criterion="entropy", n_jobs=cpu), clsnum=cls, 
    data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls")
rf = rf("b278", nfolds=10)
collector.prop_fp(rf, data_name,"b278", "k-fold")

rf = WithAnotherExperiment(
    clf=RandomForestClassifier(n_estimators=500, criterion="entropy", n_jobs=cpu), 
    clsnum=cls, data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls")
rf = rf("b278")
for result in rf:
    collector.prop_fp(result, data_name,"b278", result.test_name)
    
rf = WithAnotherExperiment(
    clf=RandomForestClassifier(n_estimators=500, criterion="entropy", n_jobs=cpu), 
    clsnum=cls, data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls")
rf = rf("b261")
for result in rf:
    collector.prop_fp(result, data_name,"b261", result.test_name)
    
rf = WithAnotherExperiment(
    clf=RandomForestClassifier(n_estimators=500, criterion="entropy", n_jobs=cpu), 
    clsnum=cls, data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls")
rf = rf(["b278", "b261"])
for result in rf:
    collector.prop_fp(result, data_name,"b261 + b278", result.test_name)
    
rf = WithAnotherExperiment(
    clf=RandomForestClassifier(n_estimators=500, criterion="entropy", n_jobs=cpu), 
    clsnum=cls, data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls")
rf = rf(["b278", "b261", "b264"])
for result in rf:
    collector.prop_fp(result, data_name,"b261 + b264 + b278", result.test_name)

## Medium

%%time
data = data_mid
data_name = "mid"

rf = KFoldExperiment(
    clf=RandomForestClassifier(n_estimators=500, criterion="entropy", n_jobs=cpu), clsnum=cls, 
    data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls")
rf = rf("b278", nfolds=10)
collector.prop_fp(rf, data_name,"b278", "k-fold")

rf = WithAnotherExperiment(
    clf=RandomForestClassifier(n_estimators=500, criterion="entropy", n_jobs=cpu), 
    clsnum=cls, data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls")
rf = rf("b278")
for result in rf:
    collector.prop_fp(result, data_name,"b278", result.test_name)
    
rf = WithAnotherExperiment(
    clf=RandomForestClassifier(n_estimators=500, criterion="entropy", n_jobs=cpu), 
    clsnum=cls, data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls")
rf = rf("b261")
for result in rf:
    collector.prop_fp(result, data_name,"b261", result.test_name)
    
rf = WithAnotherExperiment(
    clf=RandomForestClassifier(n_estimators=500, criterion="entropy", n_jobs=cpu), 
    clsnum=cls, data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls")
rf = rf(["b278", "b261"])
for result in rf:
    collector.prop_fp(result, data_name,"b261 + b278", result.test_name)
    
rf = WithAnotherExperiment(
    clf=RandomForestClassifier(n_estimators=500, criterion="entropy", n_jobs=cpu), 
    clsnum=cls, data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls")
rf = rf(["b278", "b261", "b264"])
for result in rf:
    collector.prop_fp(result, data_name,"b261 + b264 + b278", result.test_name)

## Big

%%time
data = data_big
data_name = "big"

rf = KFoldExperiment(
    clf=RandomForestClassifier(n_estimators=500, criterion="entropy", n_jobs=cpu), clsnum=cls, 
    data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls")
rf = rf("b278", nfolds=10)
collector.prop_fp(rf, data_name,"b278", "k-fold")

rf = WithAnotherExperiment(
    clf=RandomForestClassifier(n_estimators=500, criterion="entropy", n_jobs=cpu), 
    clsnum=cls, data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls")
rf = rf("b278")
for result in rf:
    collector.prop_fp(result, data_name,"b278", result.test_name)
    
rf = WithAnotherExperiment(
    clf=RandomForestClassifier(n_estimators=500, criterion="entropy", n_jobs=cpu), 
    clsnum=cls, data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls")
rf = rf("b261")
for result in rf:
    collector.prop_fp(result, data_name,"b261", result.test_name)
    
rf = WithAnotherExperiment(
    clf=RandomForestClassifier(n_estimators=500, criterion="entropy", n_jobs=cpu), 
    clsnum=cls, data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls")
rf = rf(["b278", "b261"])
for result in rf:
    collector.prop_fp(result, data_name,"b261 + b278", result.test_name)
    
rf = WithAnotherExperiment(
    clf=RandomForestClassifier(n_estimators=500, criterion="entropy", n_jobs=cpu), 
    clsnum=cls, data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls")
rf = rf(["b278", "b261", "b264"])
for result in rf:
    collector.prop_fp(result, data_name,"b261 + b264 + b278", result.test_name)

np.save("coso.npy", [collector])

collector = np.load("coso.npy")[0]

rows = [
    ["b278", [
        "k-fold",
        "b261",
        "b262",
        "b263",
        "b264"]],
    ["b261", [
        "b262",
        "b263",
        "b264",
        "b278"]],
    ["b261 + b278", [
        "b262",
        "b263",
        "b264"]],
    ["b261 + b264 + b278", [
        "b262",
        "b263"]]
]


collector.r

columns = {}
for size in "small mid big".split():
    props = []
    for train, tests in rows:
        for test in tests:
            key = (size, train, test)
            prop = collector.r[key][0]
            props.append(prop)
    columns[size] = props
    
trainc, testc = [], []
for train, tests in rows:
    for test in tests:
        trainc.append(train)
        testc.append(test)
        train = ""
columns.update(train=trainc, test=testc)
    

print pd.DataFrame(columns)["train test small mid big".split()].to_latex(index=False)
