==> 07_scaler-o3o4vZ.ipynb <==
# Escalado

import pickle

import numpy as np

import pandas as pd

% matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns
# sns.set()

from sklearn.preprocessing import StandardScaler

s2_5k = pd.read_pickle("data/o3o4vZ/s2_5k.pkl.bz2")
s5k = pd.read_pickle("data/o3o4vZ/s5k.pkl.bz2")
s20k = pd.read_pickle("data/o3o4vZ/s20k.pkl.bz2")
sALL = pd.read_pickle("data/o3o4vZ/sampleALL_b261_b278.pkl.bz2")

## Removemos todo lo que no sea RRLyrae o desconocido

def filter_only_rr_unk(df):
    flt = (df.vs_type == "") | df.vs_type.str.startswith('RRLyr-')
    return df[flt]

s2_5k = filter_only_rr_unk(s2_5k)
s5k = filter_only_rr_unk(s5k)
s20k = filter_only_rr_unk(s20k)
sALL = filter_only_rr_unk(sALL)

## Removemos los features que no queremos


to_drop = [
    'Gskew', # has nan in the rlyrae stars
    "scls_h", "scls_j", "scls_k",  # no nos sirve
    "AndersonDarling", "StetsonJ", "StetsonK"
] + [s for s in s20k.columns if s.startswith("Freq2_") or s.startswith("Freq3_") ] # only the first period is important
to_keep = [c for c in s20k.columns if c not in to_drop]

s20k = s20k[to_keep]
s5k = s5k[to_keep]
s2_5k = s2_5k[to_keep]
sALL = sALL[to_keep]

s20k["id"].apply(lambda i: "b" + str(i)[1:4]).unique()

Removemos filas que tengan un nan en `period_fit` pero antes nos fijamos que ninguna sea una RR-Lyrae

print s20k[s20k.Period_fit.isnull()].vs_type.unique()
print s5k[s5k.Period_fit.isnull()].vs_type.unique()
print s2_5k[s2_5k.Period_fit.isnull()].vs_type.unique()
print sALL[sALL.Period_fit.isnull()].vs_type.unique()

sALL.columns[sALL.isnull().any()]

s20k = s20k.dropna()
s5k = s5k.dropna()
s2_5k = s2_5k.dropna()
sALL = sALL.dropna()

print s20k["id"].apply(lambda i: "b" + str(i)[1:4]).unique()
print sALL["id"].apply(lambda i: "b" + str(i)[1:4]).unique()

Evaluamos que columnas tienen un valor infinito

for x in s20k.columns:
    if s20k[x].dtype == object:
        continue
    if np.isinf(s20k[x].values).sum() + np.isinf(s5k[x].values).sum() + np.isinf(s2_5k[x].values).sum() + np.isinf(sALL[x].values).sum():
        print x

Como period_fit es un feature que me interesa, verificamos que cantidad de filas son las afectadas
y si hay alguna con RRLyraes

print s2_5k[np.isinf(s2_5k.Period_fit.values)].shape
print s5k[np.isinf(s5k.Period_fit.values)].shape
print s20k[np.isinf(s20k.Period_fit.values)].shape
print sALL[np.isinf(sALL.Period_fit.values)].shape

print s2_5k[np.isinf(s2_5k.Period_fit.values)].vs_type.unique()
print s5k[np.isinf(s5k.Period_fit.values)].vs_type.unique()
print s20k[np.isinf(s20k.Period_fit.values)].vs_type.unique()
print sALL[np.isinf(sALL.Period_fit.values)].vs_type.unique()

Son pocas filas y no hay rrlyraes... las borramos

s20k = s20k[~np.isinf(s20k.Period_fit.values)]
s5k = s5k[~np.isinf(s5k.Period_fit.values)]
s2_5k = s2_5k[~np.isinf(s2_5k.Period_fit.values)]
sALL = sALL[~np.isinf(sALL.Period_fit.values)]

ahora almacenamos tod esto limpio para futuros usos

s20k.to_pickle("data/o3o4vZ/nonull/s20k.pkl.bz2", compression="bz2")
s5k.to_pickle("data/o3o4vZ/nonull/s5k.pkl.bz2", compression="bz2")
s2_5k.to_pickle("data/o3o4vZ/nonull/s2_5k.pkl.bz2", compression="bz2")
sALL.to_pickle("data/o3o4vZ/nonull/sALL.pkl.bz2", compression="bz2")

## Ahora normalizamos

no_features = ["id", "vs_catalog", "vs_type", "ra_k", "dec_k", ] 
X_columns = [c for c in s20k.columns if c not in no_features]

splited = {idx+1: list(s) for idx, s in enumerate(np.array_split(X_columns, 19))}
maxs = max(map(len, splited.values()))
for v in splited.values():
    while len(v) < maxs:
        v.append("")

scaler_20k = StandardScaler()
norm_s20k = s20k.copy()
norm_s20k[X_columns] = scaler_20k.fit_transform(s20k[X_columns])
pickle.dump(scaler_20k, open("data/o3o4vZ/scalers/scaler_20k.pkl", "wb"))
norm_s20k.to_pickle("data/o3o4vZ/scaled/s20k.pkl.bz2", compression="bz2")

scaler_5k = StandardScaler()
norm_s5k = s5k.copy()
norm_s5k[X_columns] = scaler_5k.fit_transform(s5k[X_columns])
pickle.dump(scaler_5k, open("data/o3o4vZ/scalers/scaler_5k.pkl", "wb"))
norm_s5k.to_pickle("data/o3o4vZ/scaled/s5k.pkl.bz2", compression="bz2")

scaler_2_5k = StandardScaler()
norm_s2_5k = s2_5k.copy()
norm_s2_5k[X_columns] = scaler_2_5k.fit_transform(s2_5k[X_columns])
pickle.dump(scaler_2_5k, open("data/o3o4vZ/scalers/scaler_2_5k.pkl", "wb"))
norm_s2_5k.to_pickle("data/o3o4vZ/scaled/s2_5k.pkl.bz2", compression="bz2")

cosos = ['b220',
 'b234',
 'b247',
 'b248',
 'b261',
 'b262',
 'b263',
 'b264',
 'b277',
 'b278',
 'b396']
rows = {k: {"Tile": k} for k in cosos}
    
for nombre, s in zip(["Chica", "Mediana", "Grande"] , [s2_5k,s5k,s20k]):
    s = s.copy()
    s["tile"] = s["id"].apply(lambda i: "b" + str(i)[1:4])
    grouped = s.groupby("tile")
    data = {k: grouped.get_group(k).copy() for k in grouped.groups.keys()}
    for tile, df in data.items():
        row = rows[tile]
        row.update({
            nombre : len(df[df.vs_type.str.contains("")]),
            "RR-Lyrae": len(df[df.vs_type.str.startswith("RRLyr")]),
        })
        



s = s20k.copy()
s["tile"] = s["id"].apply(lambda i: "b" + str(i)[1:4])

s.tile.unique()

for r in rows.values():
    print r

print pd.DataFrame(
    [v for k, v in sorted(rows.items())]
)[
    "Tile Chica Mediana Grande RR-Lyrae".split()
].sort_values("Tile").to_latex(index=False)
