==> 08_all_vs_all_vs-Copy1.ipynb <==
# All vs All

import pickle

import numpy as np

import pandas as pd

% matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns
# sns.set()

import joblib

from sklearn.svm import SVC
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.neighbors import KNeighborsClassifier

from sklearn.preprocessing import StandardScaler

from libs.container import Container
from libs.display import d
from libs.nearest import nearest
from libs.experiment import KFoldExperiment, WithAnotherExperiment, roc, metrics
from libs.precstar import  prec_star

sample = pd.read_pickle("data/o3o4vZ/scaled/s2_5k.pkl.bz2")
sample["tile"] = sample["id"].apply(lambda i: "b" + str(i)[1:4])
sample["cls"] = sample.vs_type.apply(lambda x: 0 if x == "" else 1)

no_features = ["id", "vs_catalog", "vs_type", "ra_k", "dec_k", "tile", "cls"] 
X_columns = [c for c in sample.columns if c not in no_features]

grouped = sample.groupby("tile")
data_small = Container({k: grouped.get_group(k).copy() for k in grouped.groups.keys()})

del grouped, sample

sample = pd.read_pickle("data/o3o4vZ/scaled/s5k.pkl.bz2")
sample["tile"] = sample["id"].apply(lambda i: "b" + str(i)[1:4])
sample["cls"] = sample.vs_type.apply(lambda x: 0 if x == "" else 1)

grouped = sample.groupby("tile")
data_mid = Container({k: grouped.get_group(k).copy() for k in grouped.groups.keys()})

del grouped, sample

sample = pd.read_pickle("data/o3o4vZ/scaled/s20k.pkl.bz2")
sample["tile"] = sample["id"].apply(lambda i: "b" + str(i)[1:4])
sample["cls"] = sample.vs_type.apply(lambda x: 0 if x == "" else 1)

grouped = sample.groupby("tile")
data_big = Container({k: grouped.get_group(k).copy() for k in grouped.groups.keys()})

del grouped, sample

## Features

results = {}
cpu = joblib.cpu_count()

cls = {0:0, 1:1}

d(X_columns)

SIZES = {
    'b220': 211850,
    'b234': 297302,
    'b247': 414497,
    'b248': 426369,
    'b261': 575075,
    'b262': 591770,
    'b263': 585661,
    'b264': 614967,
    'b277': 753146,
    'b278': 781612,
    'b396': 494646}

SP = .1

def get_prec_star(r, tile_name):
    rs = SIZES[tile_name]
    return prec_star(r.y_test, r.probabilities[:,1], r.test_size, rs)

def get_metrics(kf, vss, train_name):
    # kfold correction
    
    pstar = get_prec_star(kf, train_name)
    
    idx = nearest(pstar, SP)
    precs, recs, curve = kf.prec_rec_curve
    kfold_prec = precs[idx]
    kfold_recall = recs[idx]
    
    m = Container(
        kfold=(kfold_prec, kfold_recall), vss=Container())
    
    for vs in vss:
        pstar = get_prec_star(vs, vs.test_name)
        idx = nearest(pstar, SP)
        
        precs, recs, curve = vs.prec_rec_curve
        prec = precs[idx]
        recall = recs[idx]
        m.vss[vs.test_name] = (prec, recall)
    
    return m


def run(train, data):
    print ">>>> Kfolding {} <<<<".format(train)
    kf = KFoldExperiment(
        clf=RandomForestClassifier(n_estimators=500, criterion="entropy", n_jobs=cpu), clsnum=cls, 
        data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls", verbose=False)
    kf = kf(train, nfolds=10)
    
    print ">>>> Vs {}<<<<".format(train)
    vs = WithAnotherExperiment(
        clf=RandomForestClassifier(n_estimators=500, criterion="entropy"), verbose=False, 
        clsnum=cls, data=data, pcls=1, ncls=0, X_columns=X_columns, y_column="cls")
    vs = vs(train)
    
    return train, get_metrics(kf=kf, vss=vs, train_name=train)

cpu = joblib.cpu_count()
print cpu

with joblib.Parallel(n_jobs=cpu) as jobs:
    result = jobs(
        joblib.delayed(run)(k, data_small)
        for k in sorted(data_small.keys()))
results["small"] = dict(result)

with joblib.Parallel(n_jobs=cpu) as jobs:
    result = jobs(
        joblib.delayed(run)(k, data_mid)
        for k in sorted(data_mid.keys()))
results["mid"] = dict(result)

with joblib.Parallel(n_jobs=cpu) as jobs:
    result = jobs(
        joblib.delayed(run)(k, data_big)
        for k in sorted(data_big.keys()))
results["big"] = dict(result)

np.save("data/o3o4vZ/all_vs_all_vs/results.npy", [results])

for k, data in sorted(data_big.items()):
    print k, len(data[data.vs_type!=""])
